Fixing selecting bar patients...
Add log4j2 dependency in pom.xml.
Add log4j2.yaml to be able to use the dependency
Resolving POST call in ReferentialPositionController.
The file application.properties was modified in order to get a local DB.

+++ src/main/java/com/application/aled/repository/ReferentialPositionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.application.aled.entity.ReferentialPosition;
 
 @Repository
-public interface ReferentialPositionRepository extends CrudRepository<ReferentialPosition, Long> {
+public interface ReferentialPositionRepository extends CrudRepository<ReferentialPosition, Integer> {
 
 }
